Fix poll answer images not being displayed
After updating Foundation in commit 58071fd6, the orbit slider stopped
working properly. That's because the `.orbit-slide` elements now use a
`position: absolute` rule, and so our rule for `.orbit-container`
elements making their height 100% (which we added in order to add images
with different heights, which Foundation's orbit doesn't seem to handle
properly) makes them have a height of 0px, since now the `.orbit-slide`
elements are not part of the document flow anymore.

Making the `.orbit-slide` elements have relative position fixes this
issue, but introduces a different one, producing a really bad-looking
animation when changing a slide.

So we're disabling the animation as well in order to avoid this jump.

This change also fixes another issue introduced in commit 58071fd6 which
caused slide controls to stop working when changing slides back and
forth.